src/hardware/upload_server/requirements.txt
azure_storage==0.37.0
Flask==3.1.1
Requests==2.32.3
